frontend

+++ src/main/java/moneytransferapp/dto/TransactionRequest.java
@@ -1,4 +1,4 @@
 package moneytransferapp.dto;
 
-public record TransactionRequest(String fromAccount, String toAccount, int amountToTransfer) {
+public record TransactionRequest(String fromAccount, String toAccount, double amountToTransfer) {
 }
